cloud_functions/profile_manager_v2/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.14.0
firebase-admin==6.4.0
google-cloud-storage==2.13.0
google-genai>=1.0.0
google-generativeai>=0.8.0
requests==2.31.0
PyMuPDF==1.24.0
python-docx==1.1.0
Pillow>=10.0.0
